Some API relocations and documentation patches
