+++ src/test/java/io/jenkins/plugins/extlogging/api/util/MockLoggingMethod.java
@@ -1,7 +1,7 @@
 package io.jenkins.plugins.extlogging.api.util;
 
 import hudson.model.Run;
-import io.jenkins.plugins.extlogging.api.impl.ExternalLoggingEventWriter;
+import io.jenkins.plugins.extlogging.api.ExternalLoggingEventWriter;
 import io.jenkins.plugins.extlogging.api.ExternalLoggingMethod;
 import jenkins.model.logging.LogBrowser;
